WL-3959: Click 'Add Section' button shows inline ck editor
I've had to put the inline editor creation in the javascript on the page
because I get an 'uncaught exception: ckeditor instance already defined on the page' error otherwise.
 otherwise.

The handler is in a js file.  new_resources.js.  It may be that this ORLIMS needs
its own js file and I'll create one when we get to that stage.