WL-3959: Add ability to create separate multiple ck editors
I'm creating a ck editor with a different name each time by including and
incrementing locationId by 1 each time.  This means that each editor is
independent and separate.

I'm also passing through locationId to the Java testing whether the section
already exists and updating it if it does (or inserting it if it doesn't).