WL-3935: Validate nested sections
There have been some issues where some
dodgy data has been getting saved to the db, for example an h3 after
an h1- which should not happen because the UI should not allow it.
So when the user does a drag and drop, we are validating the nested structure
server-side and throwing an exception if it's not right.

WL-3935: Moving validation into static model

WL-3935: Rename Validator to CitationValidator

WL-3935: Add comment ; remove citation from top level ; indentation

WL-3935: Add citation back in